Update !radar with new caching
A problem with the radar command is that it is fetched for each request, so it shows the radar at the time when it was loaded by the user, not when it was posted to slack. I've updated the service to take a timestamp and it'll use a cached image from that timestamp if available.